requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
idna==3.10
itsdangerous==2.2.0
MarkupSafe==3.0.2
narwhals==1.39.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
setuptools==80.7.1
six==1.17.0
tenacity==9.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
